move check for preload issue inside ZoomSDK.web
